requirements.txt
# Web Framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
websockets>=12.0

# LLM APIs
anthropic>=0.49.0
google-generativeai>=0.4.0

# Data Validation
pydantic>=2.6.0
jsonschema>=4.21.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0
aiofiles>=23.2.0
